Rename duke to inventoryBro in relevant class, files and package names
