+++ docs/DeveloperGuide.md
@@ -20,10 +20,10 @@
 
 ## User Stories
 
-|Version| As a ... | I want to ... | So that I can ...|
-|--------|----------|---------------|------------------|
-|v1.0|new user|see usage instructions|refer to them when I forget how to use the application|
-|v2.0|user|find a to-do item by name|locate a to-do without having to go through the entire list|
+| Version  | As a ...  | I want to ...             | So that I can ...                                           |
+|----------|-----------|---------------------------|-------------------------------------------------------------|
+| v1.0     | new user  | see usage instructions    | refer to them when I forget how to use the application      |
+| v2.0     | user      | find a to-do item by name | locate a to-do without having to go through the entire list |
 
 ## Non-Functional Requirements
