+++ src/main/java/seedu/inventorybro/InventoryBro.java
@@ -1,0 +1,13 @@
+package seedu.inventorybro;
+
+import java.util.Scanner;
+
+public class InventoryBro {
+
+    /**
+     * Main entry-point for the java.seedu.inventorybro.InventoryBro application.
+     */
+    public static void main(String[] args) {
+
+    }
+}
