+++ src/test/java/seedu/inventorybro/InventoryBroTest.java
@@ -1,10 +1,10 @@
-package seedu.duke;
+package seedu.inventorybro;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-class DukeTest {
+class InventoryBroTest {
     @Test
     public void sampleTest() {
         assertTrue(true);
